GUI/Subscriber/source/requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
paho-mqtt==1.6.1
pillow==10.2.0
pygame==2.5.2
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
